Rename and refactor charge confirmation endpoint
