+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Charges/MessageHub/ChargeConfirmationDataAvailableNotifierEndpoint.cs
@@ -14,26 +14,32 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-using GreenEnergyHub.Charges.Application.Charges.Acknowledgement;
+using GreenEnergyHub.Charges.Application.MessageHub;
+using GreenEnergyHub.Charges.Domain.AvailableChargeReceiptData;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommandAcceptedEvents;
 using GreenEnergyHub.Charges.FunctionHost.Common;
 using GreenEnergyHub.Charges.Infrastructure.Internal.ChargeCommandAccepted;
 using GreenEnergyHub.Charges.Infrastructure.Messaging;
 using Microsoft.Azure.Functions.Worker;
 
-namespace GreenEnergyHub.Charges.FunctionHost.Charges
+namespace GreenEnergyHub.Charges.FunctionHost.Charges.MessageHub
 {
-    public class ChargeConfirmationSenderEndpoint
+    /// <summary>
+    /// The function will initiate the communication with the message hub
+    /// by notifying that a charge change has been confirmed
+    /// This is the RSM-033 CIM XML 'ConfirmRequestChangeBillingMasterData'.
+    /// </summary>
+    public class ChargeConfirmationDataAvailableNotifierEndpoint
     {
-        public const string FunctionName = nameof(ChargeConfirmationSenderEndpoint);
-        private readonly IChargeConfirmationSender _chargeConfirmationSender;
+        public const string FunctionName = nameof(ChargeConfirmationDataAvailableNotifierEndpoint);
+        private readonly IAvailableDataNotifier<AvailableChargeReceiptData, ChargeCommandAcceptedEvent> _availableDataNotifier;
         private readonly MessageExtractor<ChargeCommandAcceptedContract> _messageExtractor;
 
-        public ChargeConfirmationSenderEndpoint(
-            IChargeConfirmationSender chargeConfirmationSender,
+        public ChargeConfirmationDataAvailableNotifierEndpoint(
+            IAvailableDataNotifier<AvailableChargeReceiptData, ChargeCommandAcceptedEvent> availableDataNotifier,
             MessageExtractor<ChargeCommandAcceptedContract> messageExtractor)
         {
-            _chargeConfirmationSender = chargeConfirmationSender;
+            _availableDataNotifier = availableDataNotifier;
             _messageExtractor = messageExtractor;
         }
 
@@ -46,7 +52,7 @@ namespace GreenEnergyHub.Charges.FunctionHost.Charges
             [NotNull] byte[] message)
         {
             var acceptedEvent = (ChargeCommandAcceptedEvent)await _messageExtractor.ExtractAsync(message).ConfigureAwait(false);
-            await _chargeConfirmationSender.HandleAsync(acceptedEvent).ConfigureAwait(false);
+            await _availableDataNotifier.NotifyAsync(acceptedEvent).ConfigureAwait(false);
         }
     }
 }
